New Async, Fixed loading issues, Fixed duplicate loadings, Fixed this and that, fixed a ton, work work work - Improved dynamic loading in CatalogController.kt - ... - I gave up commenting too much stuff

+++ android/src/androidTest/java/app/shosetsu/android/BackupTest.kt
@@ -83,7 +83,7 @@ class BackupTest : KodeinAware {
 					randomString, randomString,
 					Random.nextBoolean(),
 					ReadingStatus.fromInt(Random.nextInt() % 3 + 1),
-					randomInt
+					randomInt.toDouble()
 				)
 			)
 		}
